DevOps/requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.7.0
billiard==3.6.4.0
bleach==5.0.1
boto3==1.20.26
botocore==1.23.26
CacheControl==0.12.10
cachetools==4.2.4
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.45.0
django-archive==0.2.0
django-browser-reload==1.6.0
django-ckeditor==6.5.1
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==2.0.0
django-payu==0.5
django-rest-auth==0.9.5
django-storages==1.12.3
django-summernote==0.8.20.0
django-tailwind==3.4.0
djangorestframework==3.12.4
environ==1.0
fcm-django==1.0.7
firebase-admin==5.2.0
GDAL==3.5.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
jmespath==0.10.0
kombu==5.2.3
msgpack==1.0.3
mysql-connector-python==8.0.30
numpy==1.22.3
oauthlib==3.1.1
packaging==21.3
paywix==1.4.0
Pillow==9.0.0
postgis==1.0.4
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.3
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
sendgrid==6.8.3
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
toml==0.10.2
types-cryptography==3.3.23
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
yapf==0.32.0
